# apps/backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Our Haven API
  version: 0.0.0
  description: Our Haven v1 API — Parent + Provider mobile, Provider web portal,
    admin dashboard. Source of truth per ADR-0004 (§§ 1–3, 8) and ADR-0010
    (Supabase + Fly.io + Vercel platform). US-region only.
  license:
    name: Proprietary
components:
  securitySchemes:
    supabaseAccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token (HS256). Verified locally with the
        project JWT secret on every request.
...
paths:
...
  /v1/providers/me/verification:
    get:
      summary: Read the authenticated Provider's verification state + checklist facts
      tags:
        - providers
      description: Returns the current Verification state computed from per-step facts
        (email/phone/ID/screening/license) plus the raw timestamps that drive
        the design's 8-step checklist. Email + phone confirmations are mirrored
        from Supabase Auth on read so they reflect the latest session state.
      security:
        - supabaseAccessToken: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  state:
                    type: string
                    enum:
                      - unverified
                      - email-verified
                      - phone-verified
                      - id-uploaded
                      - screening-initiated
                      - screening-passed
                      - license-pending
                      - license-verified
                      - connect-pending
                      - activated
                      - rejected
                      - holding-state-not-supported
                  kind:
                    type: string
                    enum:
                      - caregiver
                      - specialist
                  residentState:
                    type: string
                    enum:
                      - AL
                      - AK
                      - AZ
                      - AR
                      - CA
                      - CO
                      - CT
                      - DE
                      - DC
                      - FL
                      - GA
                      - HI
                      - ID
                      - IL
                      - IN
                      - IA
                      - KS
                      - KY
                      - LA
                      - ME
                      - MD
                      - MA
                      - MI
                      - MN
                      - MS
                      - MO
                      - MT
                      - NE
                      - NV
                      - NH
                      - NJ
                      - NM
                      - NY
                      - NC
                      - ND
                      - OH
                      - OK
                      - OR
                      - PA
                      - RI
                      - SC
                      - SD
                      - TN
                      - TX
                      - UT
                      - VT
                      - VA
                      - WA
                      - WV
                      - WI
                      - WY
                  licenseBoardSupported:
                    type: boolean
                  facts:
                    type: object
                    properties:
                      emailConfirmedAt:
                        anyOf:
                          - type: string
                            format: date-time
                            pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                          - type: "null"
                      phoneConfirmedAt:
                        anyOf:
                          - type: string
                            format: date-time
                            pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                          - type: "null"
                      idDocUploadedAt:
                        anyOf:
                          - type: string
                            format: date-time
                            pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                          - type: "null"
                      idDocObjectPath:
                        anyOf:
                          - type: string
                          - type: "null"
                      screeningInitiatedAt:
                        anyOf:
                          - type: string
                            format: date-time
                            pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                          - type: "null"
                      screeningPassedAt:
                        anyOf:
                          - type: string
                            format: date-time
                            pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                          - type: "null"
                      licenseVerifiedAt:
                        anyOf:
                          - type: string
                            format: date-time
                            pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                          - type: "null"
                      connectAccountReadyAt:
                        anyOf:
                          - type: string
                            format: date-time
                            pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                          - type: "null"
                      connectChargesEnabled:
                        type: boolean
                      connectPayoutsEnabled:
                        type: boolean
                      rejectedAt:
                        anyOf:
                          - type: string
                            format: date-time
                            pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d(?::[0-5]\d(?:\.\d+)?)?(?:Z))$
                          - type: "null"
                      rejectionReason:
                        anyOf:
                          - type: string
                          - type: "null"
                    required:
                      - emailConfirmedAt
                      - phoneConfirmedAt
                      - idDocUploadedAt
                      - idDocObjectPath
                      - screeningInitiatedAt
                      - screeningPassedAt
                      - licenseVerifiedAt
                      - connectAccountReadyAt
                      - connectChargesEnabled
                      - connectPayoutsEnabled
                      - rejectedAt
                      - rejectionReason
                    additionalProperties: false
                required:
                  - state
                  - kind
                  - residentState
                  - licenseBoardSupported
                  - facts
                additionalProperties: false
        "401":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "403":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
        "404":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  reason:
                    type: string
                required:
                  - error
                additionalProperties: false
...
servers:
  - url: https://api.ourhaven.example
    description: Production (Fly.io iad)
  - url: http://localhost:8080
    description: Local development
...